requirements.txt
python-dateutil
clean-text[gpl]==0.6.0
pandas==2.1.1
